Remove shared Configurators
