+++ archive/progress_async/src/main/scala/uk/ac/wellcome/platform/archive/progress_async/config/ArgsConfigurator.scala
@@ -1,20 +1,53 @@
 package uk.ac.wellcome.platform.archive.progress_async.config
 
-import org.rogach.scallop.ScallopConf
-import uk.ac.wellcome.platform.archive.common.config._
+import org.rogach.scallop.{ScallopConf, ScallopOption}
+import uk.ac.wellcome.messaging.sns.SNSConfig
+import uk.ac.wellcome.messaging.sqs.SQSConfig
+import uk.ac.wellcome.monitoring.MetricsConfig
 import uk.ac.wellcome.platform.archive.common.modules._
 import uk.ac.wellcome.platform.archive.common.progress.modules.ProgressMonitorConfig
 import uk.ac.wellcome.platform.archive.progress_async.models.ProgressAsyncConfig
 import uk.ac.wellcome.storage.dynamo.DynamoConfig
 
+import scala.concurrent.duration._
+
 class ArgsConfigurator(val arguments: Seq[String])
-    extends ScallopConf(arguments)
-    with CloudWatchClientConfigurator
-    with MetricsConfigConfigurator
-    with SnsClientConfigurator
-    with SnsConfigConfigurator
-    with SqsClientConfigurator
-    with SqsConfigConfigurator {
+  extends ScallopConf(arguments) {
+
+  private val awsCloudwatchRegion = opt[String](default = Some("eu-west-1"))
+  private val awsCloudwatchEndpoint = opt[String]()
+
+  val cloudwatchClientConfig = CloudwatchClientConfig(
+    region = awsCloudwatchRegion(),
+    endpoint = awsCloudwatchEndpoint.toOption
+  )
+
+  private val metricsNamespace = opt[String](default = Some("app"))
+  private val metricsFlushIntervalSeconds =
+    opt[Int](required = true, default = Some(20))
+
+  val metricsConfig = MetricsConfig(
+    namespace = metricsNamespace(),
+    flushInterval = metricsFlushIntervalSeconds() seconds
+  )
+
+  private val awsSnsAccessKey = opt[String]()
+  private val awsSnsSecretKey = opt[String]()
+  private val awsSnsRegion = opt[String](default = Some("eu-west-1"))
+  private val awsSnsEndpoint = opt[String]()
+
+  val snsClientConfig = SnsClientConfig(
+    accessKey = awsSnsAccessKey.toOption,
+    secretKey = awsSnsSecretKey.toOption,
+    region = awsSnsRegion(),
+    endpoint = awsSnsEndpoint.toOption
+  )
+
+  private val snsTopicArn: ScallopOption[String] = opt[String](required = true)
+
+  val snsConfig = SNSConfig(
+    topicArn = snsTopicArn()
+  )
 
   val archiveProgressMonitorTableName = opt[String](required = true)
 
@@ -23,6 +56,30 @@ class ArgsConfigurator(val arguments: Seq[String])
   val archiveProgressMonitorDynamoRegion =
     opt[String](default = Some("eu-west-1"))
   val archiveProgressMonitorDynamoEndpoint = opt[String]()
+
+  private val awsSqsAccessKey = opt[String]()
+  private val awsSqsSecretKey = opt[String]()
+  private val awsSqsRegion = opt[String](default = Some("eu-west-1"))
+  private val awsSqsEndpoint = opt[String]()
+
+  val sqsClientConfig = SQSClientConfig(
+    accessKey = awsSqsAccessKey.toOption,
+    secretKey = awsSqsSecretKey.toOption,
+    region = awsSqsRegion(),
+    endpoint = awsSqsEndpoint.toOption
+  )
+
+  private val sqsQueueUrl: ScallopOption[String] = opt[String](required = true)
+  private val sqsWaitTimeSeconds = opt[Int](required = true, default = Some(20))
+  private val sqsMaxMessages = opt[Int](required = true, default = Some(10))
+  private val sqsParallelism = opt[Int](required = true, default = Some(10))
+
+  val sqsConfig = SQSConfig(
+    queueUrl = sqsQueueUrl(),
+    waitTime = sqsWaitTimeSeconds() seconds,
+    maxMessages = sqsMaxMessages(),
+    parallelism = sqsParallelism()
+  )
 
   verify()
