+++ archive/progress_http/src/main/scala/uk/ac/wellcome/platform/archive/progress_http/Router.scala
@@ -5,11 +5,8 @@ import java.util.UUID
 import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.model.headers.Location
 import com.google.inject.Inject
-import uk.ac.wellcome.platform.archive.common.config.models.HttpServerConfig
-import uk.ac.wellcome.platform.archive.common.progress.models.{
-  Progress,
-  ProgressCreateRequest
-}
+import uk.ac.wellcome.platform.archive.common.models.HttpServerConfig
+import uk.ac.wellcome.platform.archive.common.progress.models.{Progress, ProgressCreateRequest}
 import uk.ac.wellcome.platform.archive.common.progress.monitor.ProgressMonitor
 
 import scala.util.Try
