+++ archive/progress_http/src/main/scala/uk/ac/wellcome/platform/archive/progress_http/models/ProgressHttpConfig.scala
@@ -1,7 +1,7 @@
 package uk.ac.wellcome.platform.archive.progress_http.models
 
 import uk.ac.wellcome.monitoring.MetricsConfig
-import uk.ac.wellcome.platform.archive.common.config.models.HttpServerConfig
+import uk.ac.wellcome.platform.archive.common.models.HttpServerConfig
 import uk.ac.wellcome.platform.archive.common.modules._
 import uk.ac.wellcome.platform.archive.common.progress.modules.ProgressMonitorConfig
