+++ archive/progress_http/src/main/scala/uk/ac/wellcome/platform/archive/progress_http/modules/AkkaHttpApp.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
 import com.google.inject.Injector
 import grizzled.slf4j.Logging
-import uk.ac.wellcome.platform.archive.common.config.models.HttpServerConfig
+import uk.ac.wellcome.platform.archive.common.models.HttpServerConfig
 import uk.ac.wellcome.platform.archive.progress_http.Router
 
 import scala.concurrent.{ExecutionContext, Future}
