+++ archive/progress_http/src/test/scala/uk/ac/wellcome/platform/archive/progress_http/modules/TestAppConfigModule.scala
@@ -2,7 +2,7 @@ package uk.ac.wellcome.platform.archive.progress_http.modules
 
 import com.google.inject.{AbstractModule, Provides}
 import uk.ac.wellcome.monitoring.MetricsConfig
-import uk.ac.wellcome.platform.archive.common.config.models.HttpServerConfig
+import uk.ac.wellcome.platform.archive.common.models.HttpServerConfig
 import uk.ac.wellcome.platform.archive.common.modules._
 import uk.ac.wellcome.platform.archive.common.progress.modules.ProgressMonitorConfig
 import uk.ac.wellcome.platform.archive.progress_http.models.ProgressHttpConfig
